@@ -28,133 +33,157 @@
   "dependencies": {
     "@graphistry/client-api": "^2.14.0",
     "@graphistry/client-api-react": "^2.14.0",
-    "@graphistry/falcor": "~2.9.9",
+    "@graphistry/falcor": "^2.9.11",
     "@graphistry/falcor-json-graph": "^2.9.6",
-    "@graphistry/falcor-model-rxjs": "~2.9.9",
+    "@graphistry/falcor-model-rxjs": "^2.9.11",
     "@graphistry/falcor-path-syntax": "^2.9.6",
     "@graphistry/falcor-path-utils": "^2.9.6",
-    "@graphistry/falcor-query-syntax": "~2.9.9",
-    "@graphistry/falcor-react-redux": "~2.9.9",
-    "@graphistry/falcor-react-schema": "~2.9.9",
-    "@graphistry/falcor-router": "~2.9.9",
-    "@graphistry/falcor-router-saddle": "~2.9.9",
-    "@graphistry/falcor-socket-datasource": "~2.9.9",
+    "@graphistry/falcor-query-syntax": "^2.9.11",
+    "@graphistry/falcor-react-redux": "^2.9.11",
+    "@graphistry/falcor-react-schema": "^2.9.11",
+    "@graphistry/falcor-router": "^2.9.11",
+    "@graphistry/falcor-router-saddle": "^2.9.11",
+    "@graphistry/falcor-socket-datasource": "^2.9.11",
     "@graphistry/react-split-pane": "^0.1.63",
-    "bcrypt": "1.0.2",
+    "bcrypt": "1.0.3",
     "better-react-spinkit": "2.0.0-4",
     "bootstrap": "3.3.7",
-    "bunyan": "1.8.5",
+    "bunyan": "1.8.12",
     "classnames": "2.2.5",
     "colorbrewer": "1.0.0",
-    "compression": "1.6.2",
-    "convict": "2.0.0",
-    "create-react-class": "~15.6.0",
-    "dataframe-js": "1.2.2",
-    "engine.io": "3.1.0",
-    "engine.io-client": "3.1.1",
-    "express": "4.14.0",
+    "compression": "1.7.1",
+    "convict": "4.0.0",
+    "cookie-parser": "1.4.3",
+    "create-react-class": "15.6.2",
+    "dataframe-js": "1.2.4",
+    "engine.io": "3.1.2",
+    "engine.io-client": "3.1.2",
+    "express": "4.16.0",
     "falcor-express": "0.1.4",
-    "flat": "2.0.1",
-    "glob": "7.1.1",
-    "history": "4.5.0",
+    "flat": "4.0.0",
+    "glob": "7.1.2",
+    "helmet": "3.8.2",
+    "history": "4.7.2",
+    "hpp": "0.2.2",
+    "json-stable-stringify": "1.0.1",
     "mkdirp": "0.5.1",
-    "moment": "2.17.0",
+    "moment": "2.18.1",
     "moment-timezone": "0.5.13",
     "neo4j-driver": "1.4.0",
     "node-whois": "2.1.3",
-    "object-hash": "1.1.5",
+    "object-hash": "1.1.8",
     "parse-whois": "1.0.0",
-    "passport": "0.3.2",
+    "passport": "0.4.0",
     "passport-http": "0.3.0",
     "querystring": "0.2.0",
-    "rc-switch": "1.4.3",
-    "react": "15.6.1",
-    "react-addons-shallow-compare": "15.6.0",
-    "react-bootstrap": "0.30.3",
-    "react-bootstrap-table": "2.7.1",
+    "raf": "3.3.2",
+    "rc-switch": "1.5.3",
+    "react": "16.0.0",
+    "react-addons-shallow-compare": "15.6.2",
+    "react-bootstrap": "0.31.3",
+    "react-bootstrap-table": "4.0.5",
     "react-bootstrap-timezone-picker": "1.0.11",
-    "react-dates": "12.1.2",
-    "react-dom": "15.6.1",
-    "react-fontawesome": "1.5.0",
-    "react-redux": "5.0.2",
-    "react-sanfona": "0.1.0",
-    "react-select": "1.0.0-rc.2",
-    "react-tag-input": "4.6.1",
-    "recompose": "0.25.0",
-    "redux": "3.6.0",
-    "redux-observable": "0.14.1",
-    "request": "2.79.0",
-    "rxjs": "5.4.2",
+    "react-dates": "12.6.0",
+    "react-dom": "16.0.0",
+    "react-fontawesome": "1.6.1",
+    "react-redux": "5.0.6",
+    "react-sanfona": "0.4.0",
+    "react-select": "1.0.0-rc.10",
+    "react-tag-input": "4.7.2",
+    "recompose": "0.25.1",
+    "redbox-react": "1.5.0",
+    "redux": "3.7.2",
+    "redux-observable": "0.16.0",
+    "request": "2.83.0",
+    "rxjs": "5.4.3",
     "simpleflakes": "1.2.3",
     "socket.io": "2.0.3",
     "socket.io-client": "2.0.3",
-    "source-map-support": "0.4.2",
-    "splunk-sdk": "1.8.0",
-    "sprintf-js": "~1.1.1",
-    "string-hash": "1.1.0",
+    "source-map-support": "0.4.18",
+    "splunk-sdk": "1.8.2",
+    "sprintf-js": "1.1.1",
+    "string-hash": "1.1.3",
     "string-template": "1.0.0",
-    "time-input": "1.4.0",
+    "time-input": "1.5.1",
     "underscore": "1.8.3",
-    "uws": "0.14.5",
-    "verror": "1.9.0"
+    "uws": "8.14.1",
+    "verror": "1.10.0"
   },
   "devDependencies": {
-    "assets-webpack-plugin": "3.5.0",
-    "autoprefixer": "6.6.1",
-    "babel-cli": "6.18.0",
-    "babel-core": "6.21.0",
-    "babel-loader": "6.2.10",
-    "babel-plugin-transform-runtime": "6.15.0",
-    "babel-preset-es2015": "6.18.0",
-    "babel-preset-node5": "12.0.0",
-    "babel-preset-react": "6.16.0",
-    "babel-preset-stage-0": "6.16.0",
-    "babel-runtime": "6.20.0",
-    "chai": "3.5.0",
-    "chalk": "1.1.3",
-    "cross-env": "4.0.0",
-    "css-loader": "0.26.1",
-    "eslint": "3.12.2",
-    "eslint-plugin-mocha": "4.8.0",
-    "eslint-plugin-react": "6.8.0",
-    "extract-text-webpack-plugin": "2.0.0-beta.3",
+    "assets-webpack-plugin": "3.5.1",
+    "autoprefixer": "7.1.4",
+    "babel-core": "6.26.0",
+    "babel-loader": "7.1.2",
+    "babel-plugin-dynamic-import-node": "1.1.0",
+    "babel-plugin-istanbul": "4.1.5",
+    "babel-plugin-transform-imports": "1.4.1",
+    "babel-plugin-transform-react-constant-elements": "6.23.0",
+    "babel-plugin-transform-react-remove-prop-types": "0.4.8",
+    "babel-plugin-transform-runtime": "6.23.0",
+    "babel-polyfill": "6.26.0",
+    "babel-preset-es2015": "6.24.1",
+    "babel-preset-react": "6.24.1",
+    "babel-preset-stage-0": "6.24.1",
+    "better-npm-run": "0.1.0",
+    "cache-loader": "1.0.3",
+    "chai": "4.1.2",
+    "chalk": "2.1.0",
+    "copy-webpack-plugin": "4.0.1",
+    "coveralls": "3.0.0",
+    "cross-env": "5.0.5",
+    "css-loader": "0.28.7",
+    "eslint": "4.7.2",
+    "eslint-plugin-mocha": "4.11.0",
+    "eslint-plugin-react": "7.4.0",
+    "extract-text-webpack-plugin": "3.0.0",
     "favicons-webpack-plugin": "0.0.7",
-    "file-loader": "0.9.0",
+    "file-loader": "0.11.2",
     "font-awesome": "4.7.0",
     "font-awesome-webpack": "0.0.4",
-    "json-loader": "0.5.4",
+    "friendly-errors-webpack-plugin": "1.6.1",
+    "gauge-webpack-plugin": "0.1.0",
+    "happypack": "4.0.0-beta.5",
+    "imports-loader": "0.7.1",
+    "json-loader": "0.5.7",
     "less": "2.7.2",
-    "less-loader": "2.2.3",
-    "mocha": "3.2.0",
-    "mocha-jenkins-reporter": "0.3.5",
-    "node-sass": "4.1.1",
-    "npm-install-webpack-plugin": "4.0.4",
-    "postcss-apply": "0.4.0",
-    "postcss-css-variables": "0.6.0",
-    "postcss-cssnext": "2.9.0",
+    "less-loader": "4.0.5",
+    "less-plugin-clean-css": "1.5.1",
+    "mocha": "3.5.3",
+    "mocha-jenkins-reporter": "0.3.9",
+    "mocha-webpack": "0.7.0",
+    "nodemon": "1.12.1",
+    "npm-run-all": "4.1.1",
+    "null-loader": "0.1.1",
+    "optimize-css-assets-webpack-plugin": "3.2.0",
+    "postcss": "6.0.12",
+    "postcss-apply": "0.8.0",
+    "postcss-css-variables": "0.8.0",
+    "postcss-cssnext": "3.0.2",
     "postcss-font-awesome": "0.3.4",
-    "postcss-import": "9.0.0",
-    "postcss-loader": "1.2.1",
-    "postcss-nesting": "2.3.1",
-    "pre-git": "3.14.0",
-    "progress-bar-webpack-plugin": "1.9.2",
-    "redux-devtools": "3.3.2",
-    "redux-devtools-dock-monitor": "1.1.1",
-    "redux-devtools-inspector": "0.10.0",
-    "redux-logger": "2.7.4",
-    "sass-loader": "4.1.1",
-    "shelljs": "0.7.4",
-    "stats-webpack-plugin": "0.4.2",
-    "string-replace-webpack-plugin": "0.0.4",
-    "style-loader": "0.13.1",
-    "tree-kill": "1.1.0",
-    "url-loader": "0.5.7",
-    "webpack": "2.1.0-beta.21",
-    "webpack-dashboard": "0.2.1",
-    "webpack-hot-middleware": "2.15.0",
-    "webpack-node-externals": "1.5.4",
-    "webpack-shell-plugin": "0.5.0",
-    "webpack-visualizer-plugin": "0.1.6"
+    "postcss-import": "11.0.0",
+    "postcss-loader": "2.0.6",
+    "postcss-nesting": "4.2.1",
+    "pre-git": "3.15.3",
+    "react-hot-loader": "3.0.0-beta.6",
+    "react-test-renderer": "16.0.0",
+    "redux-devtools": "3.4.0",
+    "redux-devtools-dock-monitor": "1.1.2",
+    "redux-devtools-inspector": "0.11.3",
+    "redux-logger": "3.0.6",
+    "shelljs": "0.7.8",
+    "sinon": "2.2.0",
+    "stats-webpack-plugin": "0.6.1",
+    "string-replace-webpack-plugin": "0.1.3",
+    "style-loader": "0.18.2",
+    "tree-kill": "1.2.0",
+    "url-loader": "0.5.9",
+    "webpack": "3.5.6",
+    "webpack-bundle-analyzer": "2.9.0",
+    "webpack-closure-compiler": "2.1.5",
+    "webpack-hot-middleware": "2.19.1",
+    "webpack-node-externals": "1.6.0",
+    "webpack-universal-middleware": "1.0.1",
+    "write-file-webpack-plugin": "4.2.0"
   },
   "config": {
     "pre-git": {
