@@ -1,13 +1,13 @@
-import logger from '../../../logger.js';
+import logger from 'pivot-shared/logger';
 const log = logger.createLogger(__filename);
-
-import conf from '../../../../server/config.js';
+const conf = global.__graphistry_convict_conf__;
 import { deriveTemplate } from './systemTemplates';
 
 
 // {id -> template}
-module.exports = conf.get('systemTemplates.pivots').reduce(
-	(mod, {template, ...settings}) => {		
+export const derivedTemplates = conf
+	.get('systemTemplates.pivots')
+	.reduce((mod, {template, ...settings}) => {
 		const { id, name } = settings;
 		try {
 			log.info('Adding system pivot', {id, name, 'base': template});
