@@ -1,2 +1,2 @@
-export * from './render';
 export * from './authenticate';
+export * from './HealthChecker';
