@@ -4,7 +4,7 @@ import { logErrorWithCode } from 'pivot-shared/util';
 import { $pathValue } from '@graphistry/falcor-json-graph';
 
 import VError from 'verror';
-import logger from 'pivot-shared/logger.js';
+import logger from 'pivot-shared/logger';
 const log = logger.createLogger(__filename);
 
 export default withSchema((QL, { get, set }, services) => {
