@@ -7,7 +7,7 @@ import {
 import { Observable } from 'rxjs';
 import _ from 'underscore';
 import { combineEpics } from 'redux-observable';
-import logger from '../logger.js';
+import logger from 'pivot-shared/logger';
 const log = logger.createLogger(__filename);
 
 
