@@ -1,5 +1,5 @@
 export * from './splunk.js';
-export * from './neo4j.js';
+// export * from './neo4j.js';
 export * from './whois.js';
 export * from './http.js';
 
