@@ -7,7 +7,7 @@ import {
     cloneInvestigationModel,
     clonePivotModel
 } from '../models';
-import logger from '../logger.js';
+import logger from 'pivot-shared/logger';
 const log = logger.createLogger(__filename);
 
 
