@@ -2,7 +2,7 @@ import { Observable } from 'rxjs';
 import stringhash from 'string-hash';
 import request from 'request';
 import { PivotTemplate } from './template.js';
-import logger from '../../../shared/logger.js';
+import logger from 'pivot-shared/logger';
 const log = logger.createLogger(__filename);
 
 
