@@ -1,5 +1,5 @@
 import { SplunkPivot } from '../splunkPivot.js';
-import logger from '../../../../logger.js';
+import logger from 'pivot-shared/logger';
 
 const log = logger.createLogger(__filename);
 
